New Modified Cristall ball Pdf function.
1. Automaticaly combined fit for list of data described by suffix string
--suffix=KK,UU More then 2 suffixed are supported, and 1 as well.

2. New Modified Cristall ball function.  Now I used ranges instead of
staple points.  This ranges are for left and right tail and sum of them
normalized on distance between edge and resonance position. There is
"calibration" freedom exists in simultaniously range scaling which
is descarder by normalization. To suppress this freadom I fix
first parameter of the gaussiang range to 2.

3. Write own background Pdf function. But dont use it.

4. Polinomial background coefficient c1 shoud be limited by some range
using condition of positive number of event.